Removed besu.json entries which cause issues (presumably because there was no ETC support prior to Atlantis?) (#11)
This JSON file matches the one shipped with HL Besu.

Removed the "a.k.a. Pantheon".  The new name is well known now, and we want NOT to be talking about Pantheon.
"Parity-Ethereum" vs "Parity Ethereum"

Is there any value in maintaining the copies of the JSON files in this repo after built-in config support for Mordor has been added to each of these clients, and our example command-lines are NOT using these local files?